i18n-calypso-cli: rename i18n.js to index.js to make it easier importable
