+++ packages/i18n-calypso/cli/cli.js
@@ -13,7 +13,7 @@ const program = require( 'commander' );
 /**
  * Internal dependencies
  */
-const i18n = require( './i18n' );
+const i18n = require( '.' );
 
 function collect( val, memo ) {
 	memo.push( val );
