+++ packages/i18n-calypso/cli/index.js
